requirements.txt
streamlit
requests
python-dotenv
tldextract
python-whois
dnspython
beautifulsoup4
matplotlib
pandas